Rough addition of Order Services functionality (Will only work placing marketOrders)
We are going to run into some issue since OrderService was built interacting with the exported methods in index.js. There is nuance in the parameters passed into the algodexAPi depending on the specific order type. Many ways to handle it but I think it would be good to discuss tomorrow