Addressed edgecases where takerTransObject returns undefined
Happening now because I do not have a proper mock for sell orders and am just rewriting buyOrder to have type:"sell". takerOrderObject is returning undefined because continueStrucutreLoop fails on first iteration due to asaBalance being 0. Something that the frontEnd should catch if placing a sell order but regardless good to control for that.