🐛 updated get wallet amount to return 0 if not opted in
This only happens in rare cases where an account has algo but has not opted into a single asset and is trying to purchase an ASA without first opting in.
If the account has opted into any asset it works as expected, but if the account has not opted into any assets, then accountInfo will not have an asset property and we end up performing a `.filter` on undefined. By checking if the taker is missing the asset property and returning early, we avoid this edgecase.